+++ dto/src/main/java/ca/on/oicr/gsi/runscanner/dto/ultima/MetadataAnalysisFile.java
@@ -1,0 +1,6 @@
+package ca.on.oicr.gsi.runscanner.dto.ultima;
+
+import ca.on.oicr.gsi.runscanner.dto.AnalysisFile;
+
+// Represents a non-CRAM file (e.g. index, JSON, log) produced alongside an Ultima CRAM workflow.
+public class MetadataAnalysisFile extends AnalysisFile {}